Update automattic.com to use HTTPS
On automattic.com any non-HTTPS requests are redirected to the HTTPS version.  Because of that, the link to http://automattic.com/about/ is slower than it needs to be.  Updating the link to use https://automattic.com/about/ is both more correct and a faster experience for those that click on it.